refactor(hooks): eliminate duplication between usePatientRecords and usePractitionerRecords
- Extract shared helpers (toFhirPath, appendDateParams, resourceQueryUrl, useResourceInfiniteQuery, enrichProfileData, UseRecordsResult, noop, PAGE_SIZE) into useRecordsShared.ts
- Create useRecords() internal hook accepting UseRecordsConfig (queryKeyPrefix, skipPractitionerAuthored, enrichProfiles)
- Rewrite usePatientRecords.ts and usePractitionerRecords.ts as thin wrappers
- Fix SonarQube issues: remove unused getPractitionerInfo prop, add readonly to 8 component Props types
- Remove dead useState in patient-record-soap.tsx
- Replace redundant Boolean() calls with truthy checks (handled by refactor)
- Combine 3 duplicate tests into it.each, fix duplicate test title
- Replace '\\-' with String.raw for SonarQube compliance
